refactor: Message -> Enums로 패키지명 변경
