+++ src/main/java/bridge/Enums/GameMessage.java
@@ -1,4 +1,4 @@
-package bridge.Message;
+package bridge.Enums;
 
 public enum GameMessage {
     GAME_START("다리 건너기 게임을 시작합니다.\n"),
